Renamed container folder to view folder
